backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.28.0
supabase==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.0
redis==5.0.0
